ShopDrWebParent/ShopDrBackEnd/src/main/java/com/shopdr/admin/category/CategoryRepository.java
package com.shopdr.admin.category;

import java.util.List;

import org.springframework.data.jpa.repository.Query;
import org.springframework.data.repository.PagingAndSortingRepository;

import com.shopdr.common.entity.Category;

public interface CategoryRepository extends PagingAndSortingRepository<Category, Integer> {

	@Query("SELECT c FROM Category c WHERE c.parent.id is NULL")
	public List<Category> listRootCategories();
	
	public Category findByName(String name);
	
	public Category findByAlias(String alias);
	
	
}
